+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Improvements
 - Better DICOM multi-level detection ([#1196](../../pull/1196))
-- Added an internal field to report populated tile levels in some sources ([#1197](../../pull/1197))
+- Added an internal field to report populated tile levels in some sources ([#1197](../../pull/1197), [#1199](../../pull/1199))
 
 ### Changes
 - Change how extensions and fallback priorities interact ([#1192](../../pull/1192))